+++ sdk/messaging/azservicebus/sender.go
@@ -63,11 +63,15 @@ type SendMessageOptions struct {
 // SendMessage sends a Message to a queue or topic.
 // If the operation fails it can return an *azservicebus.Error type if the failure is actionable.
 func (s *Sender) SendMessage(ctx context.Context, message *Message, options *SendMessageOptions) error {
-	err := s.links.Retry(ctx, EventSender, "SendMessage", func(ctx context.Context, lwid *internal.LinksWithID, args *utils.RetryFnArgs) error {
-		return lwid.Sender.Send(ctx, message.toAMQPMessage())
-	}, RetryOptions(s.retryOptions))
+	return s.sendMessage(ctx, message, options)
+}
 
-	return internal.TransformError(err)
+// SendAMQPMessage sends an AMQPMessage to a queue or topic.
+// Using an AMQPMessage allows for advanced use cases, like payload encoding, as well as better
+// interoperability with pure AMQP clients.
+// If the operation fails it can return an *azservicebus.Error type if the failure is actionable.
+func (s *Sender) SendAMQPMessage(ctx context.Context, message *AMQPMessage, options *SendMessageOptions) error {
+	return s.sendMessage(ctx, message, options)
 }
 
 // SendMessageBatchOptions contains optional parameters for the SendMessageBatch function.
@@ -96,14 +100,42 @@ type ScheduleMessagesOptions struct {
 // delivered can be cancelled using `Receiver.CancelScheduleMessage(s)`
 // If the operation fails it can return an *azservicebus.Error type if the failure is actionable.
 func (s *Sender) ScheduleMessages(ctx context.Context, messages []*Message, scheduledEnqueueTime time.Time, options *ScheduleMessagesOptions) ([]int64, error) {
+	return scheduleMessages(ctx, s.links, s.retryOptions, messages, scheduledEnqueueTime)
+}
+
+// ScheduleAMQPMessagesOptions contains optional parameters for the ScheduleAMQPMessages function.
+type ScheduleAMQPMessagesOptions struct {
+	// For future expansion
+}
+
+// ScheduleAMQPMessages schedules a slice of Messages to appear on Service Bus Queue/Subscription at a later time.
+// Returns the sequence numbers of the messages that were scheduled.  Messages that haven't been
+// delivered can be cancelled using `Receiver.CancelScheduleMessage(s)`
+// If the operation fails it can return an *azservicebus.Error type if the failure is actionable.
+func (s *Sender) ScheduleAMQPMessages(ctx context.Context, messages []*AMQPMessage, scheduledEnqueueTime time.Time, options *ScheduleAMQPMessagesOptions) ([]int64, error) {
+	return scheduleMessages(ctx, s.links, s.retryOptions, messages, scheduledEnqueueTime)
+}
+
+func scheduleMessages[T amqpCompatibleMessage](ctx context.Context, links internal.AMQPLinks, retryOptions RetryOptions, messages []T, scheduledEnqueueTime time.Time) ([]int64, error) {
 	var amqpMessages []*amqp.Message
 
 	for _, m := range messages {
 		amqpMessages = append(amqpMessages, m.toAMQPMessage())
 	}
 
-	ids, err := s.scheduleAMQPMessages(ctx, amqpMessages, scheduledEnqueueTime)
-	return ids, internal.TransformError(err)
+	var sequenceNumbers []int64
+
+	err := links.Retry(ctx, EventSender, "ScheduleMessages", func(ctx context.Context, lwv *internal.LinksWithID, args *utils.RetryFnArgs) error {
+		sn, err := internal.ScheduleMessages(ctx, lwv.RPC, lwv.Sender.LinkName(), scheduledEnqueueTime, amqpMessages)
+
+		if err != nil {
+			return err
+		}
+		sequenceNumbers = sn
+		return nil
+	}, retryOptions)
+
+	return sequenceNumbers, internal.TransformError(err)
 }
 
 // MessageBatch changes
@@ -129,20 +161,12 @@ func (s *Sender) Close(ctx context.Context) error {
 	return s.links.Close(ctx, true)
 }
 
-func (s *Sender) scheduleAMQPMessages(ctx context.Context, messages []*amqp.Message, scheduledEnqueueTime time.Time) ([]int64, error) {
-	var sequenceNumbers []int64
+func (s *Sender) sendMessage(ctx context.Context, message amqpCompatibleMessage, options *SendMessageOptions) error {
+	err := s.links.Retry(ctx, EventSender, "SendMessage", func(ctx context.Context, lwid *internal.LinksWithID, args *utils.RetryFnArgs) error {
+		return lwid.Sender.Send(ctx, message.toAMQPMessage())
+	}, RetryOptions(s.retryOptions))
 
-	err := s.links.Retry(ctx, EventSender, "ScheduleMessages", func(ctx context.Context, lwv *internal.LinksWithID, args *utils.RetryFnArgs) error {
-		sn, err := internal.ScheduleMessages(ctx, lwv.RPC, lwv.Sender.LinkName(), scheduledEnqueueTime, messages)
-
-		if err != nil {
-			return err
-		}
-		sequenceNumbers = sn
-		return nil
-	}, s.retryOptions)
-
-	return sequenceNumbers, err
+	return internal.TransformError(err)
 }
 
 func (sender *Sender) createSenderLink(ctx context.Context, session amqpwrap.AMQPSession) (internal.AMQPSenderCloser, internal.AMQPReceiverCloser, error) {
@@ -178,4 +202,11 @@ func newSender(args newSenderArgs) (*Sender, error) {
 
 	sender.links = args.ns.NewAMQPLinks(args.queueOrTopic, sender.createSenderLink, internal.GetRecoveryKind)
 	return sender, nil
+}
+
+// amqpCompatibleMessage is implemented by all the messages that can be
+// converted to amqp.Message
+// Implemented by AMQPMessage, MessageBatch and Message.
+type amqpCompatibleMessage interface {
+	toAMQPMessage() *amqp.Message
 }