+++ sdk/messaging/azservicebus/sender_test.go
@@ -71,8 +71,12 @@ func Test_Sender_SendBatchOfTwo(t *testing.T) {
 	}, nil)
 	require.NoError(t, err)
 
-	err = batch.AddMessage(&Message{
-		Body: []byte("[1] message in batch"),
+	err = batch.AddAMQPMessage(&AMQPMessage{
+		Body: AMQPMessageBody{
+			Data: [][]byte{
+				[]byte("[1] message in batch"),
+			},
+		},
 	}, nil)
 	require.NoError(t, err)
 
@@ -244,6 +248,74 @@ func messageFromReceivedMessage(t *testing.T, receivedMessage *ReceivedMessage) 
 	return newMsg
 }
 
+func Test_Sender_ScheduleAMQPMessages(t *testing.T) {
+	ctx := context.Background()
+
+	client, cleanup, queueName := setupLiveTest(t, nil)
+	defer cleanup()
+
+	receiver, err := client.NewReceiverForQueue(
+		queueName, &ReceiverOptions{ReceiveMode: ReceiveModeReceiveAndDelete})
+	require.NoError(t, err)
+	defer receiver.Close(context.Background())
+
+	sender, err := client.NewSender(queueName, nil)
+	require.NoError(t, err)
+	defer sender.Close(context.Background())
+
+	now := time.Now()
+	nearFuture := now.Add(20 * time.Second)
+
+	// there are two ways to schedule a message - you can use the
+	// `ScheduleMessages` API (in which case you get a sequence number that
+	// you can use with CancelScheduledMessage(s)) or you can set the
+	// `Scheduled`
+	sequenceNumbers, err := sender.ScheduleAMQPMessages(ctx,
+		[]*AMQPMessage{
+			{Body: AMQPMessageBody{Data: [][]byte{[]byte("To the future (that will be cancelled!)")}}},
+			{Body: AMQPMessageBody{Data: [][]byte{[]byte("To the future (not cancelled)")}}},
+		},
+		nearFuture, nil)
+
+	require.NoError(t, err)
+	require.EqualValues(t, 2, len(sequenceNumbers))
+
+	peekedMsg := peekSingleMessageForTest(t, receiver)
+	require.EqualValues(t, MessageStateScheduled, peekedMsg.State)
+
+	// cancel one of the ones scheduled using `ScheduleMessages`
+	err = sender.CancelScheduledMessages(ctx, []int64{sequenceNumbers[0]}, nil)
+	require.NoError(t, err)
+
+	// this isn't a typical way of doing this, but it's possible to set the field directly
+	// rather than using the simpler ScheduleAMQPMessages
+	err = sender.SendAMQPMessage(ctx,
+		&AMQPMessage{
+			Body: AMQPMessageBody{
+				Data: [][]byte{[]byte("To the future (scheduled using the field)")},
+			},
+			MessageAnnotations: map[any]any{
+				"x-opt-scheduled-enqueue-time": &nearFuture,
+			},
+		}, nil)
+
+	require.NoError(t, err)
+
+	messages, err := receiver.ReceiveMessages(ctx, 2, nil)
+	require.NoError(t, err)
+
+	// we cancelled one of the messages so it won't get enqueued (this is the one that survived)
+	require.EqualValues(t, []string{"To the future (not cancelled)", "To the future (scheduled using the field)"}, getSortedBodies(messages))
+
+	for _, m := range messages {
+		// and the scheduled enqueue time should match what we set pretty closely.
+		diff := m.ScheduledEnqueueTime.Sub(nearFuture.UTC())
+
+		// add a little wiggle room, but the scheduled time and the time we set when we scheduled it.
+		require.LessOrEqual(t, diff, time.Second, "The requested scheduled time and the actual scheduled time should be close [%s]", m.ScheduledEnqueueTime)
+	}
+}
+
 func Test_Sender_ScheduleMessages(t *testing.T) {
 	ctx := context.Background()
 
@@ -266,7 +338,6 @@ func Test_Sender_ScheduleMessages(t *testing.T) {
 	// `ScheduleMessages` API (in which case you get a sequence number that
 	// you can use with CancelScheduledMessage(s)) or you can set the
 	// `Scheduled`
-
 	sequenceNumbers, err := sender.ScheduleMessages(ctx,
 		[]*Message{
 			{Body: []byte("To the future (that will be cancelled!)")},
@@ -415,6 +486,224 @@ func TestSender_SendMessageBatchDetach(t *testing.T) {
 	}
 
 	require.EqualValues(t, []string{"0", "1", "2", "3", "4"}, getSortedBodies(all))
+}
+
+func TestSender_SendAMQPMessage(t *testing.T) {
+	client, cleanup, queueName := setupLiveTest(t, nil)
+	defer cleanup()
+
+	sender, err := client.NewSender(queueName, nil)
+	require.NoError(t, err)
+
+	err = sender.SendAMQPMessage(context.Background(), &AMQPMessage{
+		Body: AMQPMessageBody{
+			Data: [][]byte{
+				[]byte("Hello World"),
+			},
+		},
+		ApplicationProperties: map[string]interface{}{
+			"hello": "world",
+		},
+	}, nil)
+
+	require.NoError(t, err)
+
+	receiver, err := client.NewReceiverForQueue(queueName, nil)
+	require.NoError(t, err)
+
+	messages, err := receiver.ReceiveMessages(context.Background(), 1, nil)
+	require.NoError(t, err)
+
+	require.Equal(t, "Hello World", string(messages[0].Body))
+	require.Equal(t, "world", messages[0].ApplicationProperties["hello"])
+
+	require.Equal(t, [][]byte{
+		[]byte("Hello World"),
+	}, messages[0].RawAMQPMessage.Body.Data)
+
+	require.NoError(t, receiver.CompleteMessage(context.Background(), messages[0], nil))
+}
+func TestSender_SendAMQPMessageWithMultipleByteSlicesInData(t *testing.T) {
+	client, cleanup, queueName := setupLiveTest(t, nil)
+	defer cleanup()
+
+	sender, err := client.NewSender(queueName, nil)
+	require.NoError(t, err)
+
+	err = sender.SendAMQPMessage(context.Background(), &AMQPMessage{
+		DeliveryAnnotations: map[any]any{
+			"x-opt-delivery-annotation-test": "test-value",
+		},
+		MessageAnnotations: map[any]any{
+			"x-opt-message-annotation-test": "test-value",
+		},
+		Footer: map[any]any{
+			"x-opt-footer-test": "footer-value",
+		},
+		Header: &AMQPMessageHeader{
+			// These flags are just passed through - Service Bus doesn't
+			// take any action.
+			Priority: 100,
+			Durable:  true,
+		},
+		Body: AMQPMessageBody{
+			Data: [][]byte{
+				// the defacto azservicebus.Message doesn't allow you to send multiple bytes in
+				// the .Data section.
+				[]byte("Hello World"),
+				[]byte("And another message!"),
+			}},
+		ApplicationProperties: map[string]interface{}{
+			"hello": "world",
+		},
+	}, nil)
+
+	require.NoError(t, err)
+
+	receiver, err := client.NewReceiverForQueue(queueName, nil)
+	require.NoError(t, err)
+
+	messages, err := receiver.ReceiveMessages(context.Background(), 1, nil)
+	require.NoError(t, err)
+
+	m := messages[0]
+
+	// when the Body is incompatible (in this case because of multiple Data sections)
+	// the Body is nil)
+	require.Nil(t, m.Body)
+	require.NotEmpty(t, m.RawAMQPMessage.DeliveryTag)
+
+	// kill some fields that aren't important for this comparison
+	m.RawAMQPMessage.linkName = ""
+	m.RawAMQPMessage.inner = nil
+
+	require.Equal(t, &AMQPMessage{
+		Header: &AMQPMessageHeader{
+			Priority: 100,
+			Durable:  true,
+		},
+		DeliveryAnnotations: map[any]any{
+			"x-opt-delivery-annotation-test": "test-value",
+			"x-opt-lock-token":               m.RawAMQPMessage.DeliveryAnnotations["x-opt-lock-token"],
+		},
+		MessageAnnotations: map[any]any{
+			"x-opt-message-annotation-test": "test-value",
+			"x-opt-enqueued-time":           m.RawAMQPMessage.MessageAnnotations["x-opt-enqueued-time"],
+			"x-opt-sequence-number":         m.RawAMQPMessage.MessageAnnotations["x-opt-sequence-number"],
+			"x-opt-locked-until":            m.RawAMQPMessage.MessageAnnotations["x-opt-locked-until"],
+		},
+		Footer: map[any]any{
+			"x-opt-footer-test": "footer-value",
+		},
+		Properties: &AMQPMessageProperties{
+			MessageID: m.RawAMQPMessage.Properties.MessageID,
+		},
+		Body: AMQPMessageBody{Data: [][]byte{
+			[]byte("Hello World"),
+			[]byte("And another message!"),
+		}},
+		ApplicationProperties: map[string]any{
+			"hello": "world",
+		},
+		// some items come from the service so we'll just copy them over
+		// as we're not trying to test those necessarily
+		DeliveryTag: m.RawAMQPMessage.DeliveryTag,
+	}, messages[0].RawAMQPMessage)
+
+	require.NoError(t, receiver.CompleteMessage(context.Background(), messages[0], nil))
+}
+
+func TestSender_SendAMQPMessageWithValue(t *testing.T) {
+	client, cleanup, queueName := setupLiveTest(t, nil)
+	defer cleanup()
+
+	sender, err := client.NewSender(queueName, nil)
+	require.NoError(t, err)
+
+	values := []any{
+		100.0,
+		int64(101),
+		"hello world",
+		[]byte("hello world as bytes"),
+		[]string{"a", "b", "c"},
+		nil,
+	}
+
+	for i, value := range values {
+		err = sender.SendAMQPMessage(context.Background(), &AMQPMessage{
+			Body: AMQPMessageBody{Value: value},
+			ApplicationProperties: map[string]interface{}{
+				"index": i,
+			},
+		}, nil)
+		require.NoError(t, err)
+	}
+
+	receiver, err := client.NewReceiverForQueue(queueName, nil)
+	require.NoError(t, err)
+
+	messages := receiveAll(t, receiver, len(values))
+
+	sort.Slice(messages, func(i, j int) bool {
+		left := messages[i].ApplicationProperties["index"].(int64)
+		right := messages[j].ApplicationProperties["index"].(int64)
+		return left < right
+	})
+
+	for i, msg := range messages {
+		require.Equal(t, values[i], msg.RawAMQPMessage.Body.Value)
+		require.NoError(t, receiver.CompleteMessage(context.Background(), msg, nil))
+	}
+}
+
+func TestSender_SendAMQPMessageWithSequence(t *testing.T) {
+	client, cleanup, queueName := setupLiveTest(t, nil)
+	defer cleanup()
+
+	sequences := [][][]any{
+		{
+			{int64(1), "hello"},
+			{int64(2), "world"},
+		},
+		{},
+	}
+
+	for i, seq := range sequences {
+		sender, err := client.NewSender(queueName, nil)
+		require.NoError(t, err)
+
+		err = sender.SendAMQPMessage(context.Background(), &AMQPMessage{
+			Body: AMQPMessageBody{Sequence: seq},
+			ApplicationProperties: map[string]interface{}{
+				"index": i,
+			},
+		}, nil)
+
+		require.NoError(t, err)
+	}
+
+	receiver, err := client.NewReceiverForQueue(queueName, nil)
+	require.NoError(t, err)
+
+	messages := receiveAll(t, receiver, len(sequences))
+
+	sort.Slice(messages, func(i, j int) bool {
+		left := messages[i].ApplicationProperties["index"].(int64)
+		right := messages[j].ApplicationProperties["index"].(int64)
+		return left < right
+	})
+
+	for i, msg := range messages {
+		if len(sequences[i]) == 0 {
+			// this is a bit special - when we send an empty sequence, as an optimization,
+			// no value is actually sent for the section.
+			require.Nil(t, msg.RawAMQPMessage.Body.Sequence)
+			require.NoError(t, receiver.CompleteMessage(context.Background(), messages[0], nil))
+			continue
+		}
+		require.Equal(t, sequences[i], msg.RawAMQPMessage.Body.Sequence)
+		require.NoError(t, receiver.CompleteMessage(context.Background(), messages[0], nil))
+	}
 }
 
 func getSortedBodies(messages []*ReceivedMessage) []string {