+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Release History
 
-## 1.0.2 (Unreleased)
+## 1.0.2-beta.0 (Unreleased)
 
 ### Features Added
+
+- Full access to send and receive all AMQP message properties. (#18413) 
+  - Send AMQP messages using the new `AMQPMessage` type and `Sender.SendAMQPMessage()` (AMQP messages can be added to MessageBatch's as well using MessageBatch.AddAMQPMessage()). 
+  - Access the full set of AMQP message properties when receiving using the `ReceivedMessage.RawAMQPMessage` property. 
 
 ### Breaking Changes